DRY Address and Block
